Consolidate all compat in modin._compat
