+++ modin/_compat/pandas/classes.py
@@ -11,13 +11,19 @@
 # ANY KIND, either express or implied. See the License for the specific language
 # governing permissions and limitations under the License.
 
-"""Module that houses compat functions and objects for `pandas.io.common`."""
-
-from modin.core._compat import PandasCompatVersion
+from modin._compat import PandasCompatVersion
 
 if PandasCompatVersion.CURRENT == PandasCompatVersion.PY36:
-    from .py36.pd_common import get_handle
+    from .py36 import (
+        Python36CompatibleBasePandasDataset as BasePandasDatasetCompat,
+    )
+    from .py36 import Python36CompatibleDataFrame as DataFrameCompat
+    from .py36 import Python36CompatibilitySeries as SeriesCompat
 elif PandasCompatVersion.CURRENT == PandasCompatVersion.LATEST:
-    from .latest.pd_common import get_handle
+    from .latest import (
+        LatestCompatibleBasePandasDataset as BasePandasDatasetCompat,
+    )
+    from .latest import LatestCompatibleDataFrame as DataFrameCompat
+    from .latest import LatestCompatibleSeries as SeriesCompat
 
-__all__ = ["get_handle"]
+__all__ = ["BasePandasDatasetCompat", "DataFrameCompat", "SeriesCompat"]
