+++ modin/pandas/__init__.py
@@ -225,11 +225,11 @@ from .general import (
     wide_to_long,
 )
 
-from ._compat.namespace import pivot_table
-from modin.core._compat import PandasCompatVersion
+from modin._compat.pandas.namespace import pivot_table
+from modin._compat import PandasCompatVersion
 
 if PandasCompatVersion.CURRENT != PandasCompatVersion.PY36:
-    from ._compat.namespace import Flags, Float32Dtype, Float64Dtype
+    from modin._compat.pandas.namespace import Flags, Float32Dtype, Float64Dtype
 del PandasCompatVersion
 
 from .plotting import Plotting as plotting
