+++ modin/pandas/dataframe.py
@@ -44,7 +44,7 @@ from .series import Series
 from .base import BasePandasDataset, _ATTRS_NO_LOOKUP
 from .groupby import DataFrameGroupBy
 from .accessor import CachedAccessor, SparseFrameAccessor
-from ._compat.classes import DataFrameCompat
+from modin._compat.pandas.classes import DataFrameCompat
 
 
 @_inherit_docstrings(
