+++ modin/pandas/test/test_series.py
@@ -19,7 +19,7 @@ import matplotlib
 import modin.pandas as pd
 from numpy.testing import assert_array_equal
 from pandas.core.base import SpecificationError
-from modin.core._compat import PandasCompatVersion
+from modin._compat import PandasCompatVersion
 from modin.utils import get_current_execution
 from modin.test.test_utils import warns_that_defaulting_to_pandas
 import sys
